Main PCS scraping function

query_pcs scrapes PCS data (rider profiles and results) for a given vector of rider names.

query_pcs(rider_names, seasons = NULL)

Arguments

rider_names
Character vector containing one or more rider names.

seasons
Integer vector of years ("seasons") to collect results from. Will return all years if NULL.

Value

List of two data frames (profiles and results).

The results data frame contains the following:

date: Date of race.

result: Race result.

gc_result_on_stage: If a race was a stage race, gives the rider's placement overall following the stage.

race: Race name.

distance: Race distance (km).

pointspcs: PCS points awarded.

pointsuci: UCI points awarded.

stage: Type of race (one day, stage race, etc).

rider: Rider name.

team: Team name.

The profiles data frame contains the following:

rider: Rider name.

dob: Date of birth.

Nationality

pob: Place of birth.

current_team: Current team.

weight: Weight in kg.

height: Height in m.

one_day_races: Career points awarded for success in one day races.

gc: Career points awarded for success in grand tours.

tt: Career points awarded for success in time trials.

sprint: Career points awarded for success in sprint-focused races.

climber: Career points awarded for success in climbing-focused races.

Examples

# race results for two riders in the 2021 season
query_pcs(c("Peter Sagan","Adam Yates"), seasons = 2021)

#> Peter Sagan

#> Peter Sagan 2021

#> Adam Yates

#> Adam Yates 2021

#> $profiles
#> # A tibble: 2 × 12
#>   rider   dob        nationality pob   current_team  weight height one_day_races
#>   <chr>   <date>     <chr>       <chr> <chr>          <dbl>  <dbl>         <dbl>
#> 1 Adam Y… 1992-08-07 Great Brit… ""    INEOS Grenad…     58   1.73          1662
#> 2 Peter … 1990-01-26 Slovakia    ""    BORA - hansg…     78   1.84          9134
#> # … with 4 more variables: gc <dbl>, tt <dbl>, sprint <dbl>, climber <dbl>
#> 
#> $results
#> # A tibble: 150 × 10
#>    race    stage  date                result gc_result_on_st… distance pointspcs
#>    <chr>   <chr>  <dttm>               <dbl>            <int>    <dbl>     <int>
#>  1 Itzuli… Mount… 2021-04-05 00:00:00     13               NA       NA        NA
#>  2 Itzuli… Point… 2021-04-05 00:00:00      7               NA       NA        NA
#>  3 Itzuli… Gener… 2021-04-05 00:00:00      4               NA       NA       140
#>  4 La Vue… Mount… 2021-08-14 00:00:00     28               NA       NA        NA
#>  5 La Vue… Point… 2021-08-14 00:00:00     12               NA       NA        NA
#>  6 La Vue… Gener… 2021-08-14 00:00:00      4               NA       NA       220
#>  7 UAE To… Point… 2021-02-21 00:00:00      8               NA       NA        NA
#>  8 UAE To… Gener… 2021-02-21 00:00:00      2               NA       NA       150
#>  9 Volta … Mount… 2021-03-22 00:00:00      2               NA       NA        NA
#> 10 Volta … Point… 2021-03-22 00:00:00      5               NA       NA        NA
#> # … with 140 more rows, and 3 more variables: pointsuci <int>, rider <chr>,
#> #   team <chr>
#> 

# two riders, two seasons
query_pcs(c("Peter Sagan","Adam Yates"), seasons = c(2020,2021))

#> Peter Sagan

#> Peter Sagan 2021

#> Peter Sagan 2020

#> Adam Yates

#> Adam Yates 2021

#> Adam Yates 2020

#> $profiles
#> # A tibble: 2 × 12
#>   rider   dob        nationality pob   current_team  weight height one_day_races
#>   <chr>   <date>     <chr>       <chr> <chr>          <dbl>  <dbl>         <dbl>
#> 1 Adam Y… 1992-08-07 Great Brit… ""    INEOS Grenad…     58   1.73          1662
#> 2 Peter … 1990-01-26 Slovakia    ""    BORA - hansg…     78   1.84          9134
#> # … with 4 more variables: gc <dbl>, tt <dbl>, sprint <dbl>, climber <dbl>
#> 
#> $results
#> # A tibble: 258 × 10
#>    race    stage  date                result gc_result_on_st… distance pointspcs
#>    <chr>   <chr>  <dttm>               <dbl>            <int>    <dbl>     <int>
#>  1 Itzuli… Mount… 2021-04-05 00:00:00     13               NA       NA        NA
#>  2 Itzuli… Point… 2021-04-05 00:00:00      7               NA       NA        NA
#>  3 Itzuli… Gener… 2021-04-05 00:00:00      4               NA       NA       140
#>  4 La Vue… Mount… 2021-08-14 00:00:00     28               NA       NA        NA
#>  5 La Vue… Point… 2021-08-14 00:00:00     12               NA       NA        NA
#>  6 La Vue… Gener… 2021-08-14 00:00:00      4               NA       NA       220
#>  7 UAE To… Point… 2021-02-21 00:00:00      8               NA       NA        NA
#>  8 UAE To… Gener… 2021-02-21 00:00:00      2               NA       NA       150
#>  9 Volta … Mount… 2021-03-22 00:00:00      2               NA       NA        NA
#> 10 Volta … Point… 2021-03-22 00:00:00      5               NA       NA        NA
#> # … with 248 more rows, and 3 more variables: pointsuci <int>, rider <chr>,
#> #   team <chr>
#> 

# two riders, all seasons
query_pcs(c("Peter Sagan","Adam Yates"))

#> Peter Sagan

#> Peter Sagan 2021

#> Peter Sagan 2020

#> Peter Sagan 2019

#> Peter Sagan 2018

#> Peter Sagan 2017

#> Peter Sagan 2016

#> Peter Sagan 2015

#> Peter Sagan 2014

#> Peter Sagan 2013

#> Peter Sagan 2012

#> Peter Sagan 2011

#> Peter Sagan 2010

#> Peter Sagan 2009

#> Peter Sagan 2008

#> Peter Sagan 2007

#> Adam Yates

#> Adam Yates 2021

#> Adam Yates 2020

#> Adam Yates 2019

#> Adam Yates 2018

#> Adam Yates 2017

#> Adam Yates 2016

#> Adam Yates 2015

#> Adam Yates 2014

#> Adam Yates 2013

#> Adam Yates 2012

#> Adam Yates 2011

#> Adam Yates 2010

#> Adam Yates 2009

#> $profiles
#> # A tibble: 2 × 12
#>   rider   dob        nationality pob   current_team  weight height one_day_races
#>   <chr>   <date>     <chr>       <chr> <chr>          <dbl>  <dbl>         <dbl>
#> 1 Adam Y… 1992-08-07 Great Brit… ""    INEOS Grenad…     58   1.73          1662
#> 2 Peter … 1990-01-26 Slovakia    ""    BORA - hansg…     78   1.84          9134
#> # … with 4 more variables: gc <dbl>, tt <dbl>, sprint <dbl>, climber <dbl>
#> 
#> $results
#> # A tibble: 1,882 × 10
#>    race    stage  date                result gc_result_on_st… distance pointspcs
#>    <chr>   <chr>  <dttm>               <dbl>            <int>    <dbl>     <int>
#>  1 Itzuli… Mount… 2021-04-05 00:00:00     13               NA       NA        NA
#>  2 Itzuli… Point… 2021-04-05 00:00:00      7               NA       NA        NA
#>  3 Itzuli… Gener… 2021-04-05 00:00:00      4               NA       NA       140
#>  4 La Vue… Mount… 2021-08-14 00:00:00     28               NA       NA        NA
#>  5 La Vue… Point… 2021-08-14 00:00:00     12               NA       NA        NA
#>  6 La Vue… Gener… 2021-08-14 00:00:00      4               NA       NA       220
#>  7 UAE To… Point… 2021-02-21 00:00:00      8               NA       NA        NA
#>  8 UAE To… Gener… 2021-02-21 00:00:00      2               NA       NA       150
#>  9 Volta … Mount… 2021-03-22 00:00:00      2               NA       NA        NA
#> 10 Volta … Point… 2021-03-22 00:00:00      5               NA       NA        NA
#> # … with 1,872 more rows, and 3 more variables: pointsuci <dbl>, rider <chr>,
#> #   team <chr>
#>